Fix __init__.py for compass_viewer package
